Fix https://github.com/n76/Local-GSM-Backend/issues/70
Backend was remembering the last changed position and the time that it
happened and then giving it to UnifiedNLP every time an update was
requested. This causes a large number of "Ignoring location update as
it's older than other provider." messages (even though this may be the
only backend enabled.

This change makes this back end simply report the estimated position
each time the phone state listener reports a change.

Signed-off-by: Tod Fitch <Tod@FitchDesign.com>