As report current position with current time.
As long as we have not had a change in the cell towers or signal strength
being received then our location can be assumed to be the same as our
last calculation, so if UnifiedNlp asks for an update give it the most
recent calculation with the current time. This eliminates an artifact
seen by apps where it appears that no network location from this backend
is available.

Signed-off-by: Tod Fitch <Tod@FitchDesign.com>