Add LTE cell info.
The legacyGetCellTowers() function doesn't work on my phone with two sim
cards, because it uses Mnc from first sim and cell ID from second sim and
thus it fails to find the location in database. The patch allows to use
getAllCellInfo() for LTE cells, so that the legacy function doesn't need
to be used at all.

The CellInfoLte is API 17 and the code will fallback to legacy function
if LTE data were not found in database, so it should be safe to use.